Update yum related playbooks to php8 and fixed an obscure bug
Updated the yum playbooks to php8 finally.

also fixed a bug I'm 99.99999% sure no one will ever have to deal with.  if you run the role twice with the same shortname the 2nd blog instead of getting the standard welcome to wordpress page, will load up with the theme and contents of the first blog! right domain, wrong content.

but now it checks if a database with the same name already exists.