Actually fix the obscure bug. Also trying to fix php-fpm
Lies I told myself: "Oh, it should only take a few minutes to have ansible get the list of databases, then if it detects the database it's trying to create already exists, it fails the run. (to fix obscure bug if you try to add a 2nd blog with  the same domainShort"
two.hours.later I figured it out.

trying to figure out why something I faffed with with PHP now causes 502 errors without fail. removed play to copy www.conf to the php-fpm pool folder as that's probably it.