stupidity
Letsencrypt only works if nginx is stopped. starting nginx in the task immediately before generating keys is stupid.

reverted the folder permission change because that's too slow too.

removed an errant # on cfemail in the vars file that prevented letsencrypt tasks from running (only runs the task if it's defined)